Add optional support for path helpers
These are functions named `actionPath` that take all the arguments the
route takes (so, the arguments of the action, minus the arguments to the
routes because those are just global).  It generates a `typeRestrict`
function in the where clause of each path helper that never gets used,
but which causes the type inference engine to restrict the argument
types of the path helper to the types of the actual arguments to the
action.

The helpers return Network.URI.URI, a relative URI with an absolute path
If your application does not live in the root of your domain, you may
need to take the path prefix on.  For absolute URIs you can use
(Network.URI.relativeTo <baseuri> <pathhelper>)